// payload-types.ts
export interface Page {
  id: number;
  title: string;
  slug?: string | null;
  hero?: {
    type?: ('videoBackground' | 'imageBackground' | 'none') | null;
    heading?: string | null;
    subheading?: string | null;
    videoUrl?: string | null;
    backgroundImage?: (number | null) | Media;
  };
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}
// ...
export interface Footer {
  id: number;
  quickLinks?:
    | {
        link: FooterQuickLink;
        id?: string | null;
      }[]
    | null;
  contactInfo?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  socialMediaLinks?:
    | {
        platform: 'instagram' | 'twitter' | 'facebook';
        url: string;
        id?: string | null;
      }[]
    | null;
  copyrightText?: string | null;
  updatedAt?: string | null;
  createdAt?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "FooterQuickLink".
 */
export interface FooterQuickLink {
  label: string;
  type?: ('reference' | 'custom') | null;
  reference?: (number | null) | Page;
  url?: string | null;
  newTab?: boolean | null;
}